Replace temp with query method
`movie` is the movie for the rental, which is a temp made earlier on. Temporary variables like this usually get in the way when refactoring procedural code, so I prefer to first use [Replace Temp with Query](http://refactoring.com/catalog/replaceTempWithQuery.html) to turn them into a function that I can call within any extracted code.